+++ src/Players/CECP_Mediator.cpp
@@ -158,7 +158,7 @@ void CECP_Mediator::get_clock_specs()
             auto split = String::split(response);
 
             log("moves to reset clock = " + split[1]);
-            set_reset_moves(std::stoi(split[1]));
+            set_reset_moves(std::stoul(split[1]));
 
             auto time_split = String::split(split[2], ":");
             auto game_time = 0;
